Remove and show duplicates
Return a list of unique elements - this is done by casting the list into a set, and then casting the set back to a list. Sets can only contain unique elements

List of duplicates - iterate through the original list and add each number to a dictionary. We keep track of the frequency of each element in the list. After, we can iterate through the dictionary, every key with a value higher than 1 can be added to a list of duplicates.